refactor(DuplicatePhraseManager): break 589-line component into hook + parts
- useDuplicateManager hook owns all selection/merge state and actions
- BulkActionsBar, DuplicateGroupCard, PhraseVariantRow extracted as components
- New shared ModalBase primitive (Radix Dialog + motion scaffold) replaces the
  duplicated overlay/content wrapper; reusable by the other modals
- Component becomes a folder with index.tsx orchestrator; import path unchanged

Verified: tsc clean, lint 0 errors, build succeeds. No behavior change.
